+++ .vscode/launch.json
@@ -5,8 +5,6 @@
     "version": "0.2.0",
     "configurations": [
 
-
-
         {
             "name": "Factor - Debug",
             "type": "python",
