+++ README.md
@@ -1,3 +1,15 @@
-# FlaskFactor
-An attempt at a flask POC... eventually to become an inventory management system for RPG adventuring...
+# Flask Factor
+
+*Evan Young February 2021*  
+
+## Summary  
+Flack Factor is a treasure managment system for Pathfinder RPG, it attemts to provide a place to collect treasure that our adventuring company collects, value it and distribute it. It could be given to a company member, stored, or sold by our 'Factor' a kind of merchant. The money earned from sales is returned to the company and can be distributed to the members of the company with an option for a percentage to be held in the company 'account'. 
+
+## Note
+This is a Python / Flask / sqlite etc. learning project for me so please don't use this as a good example of programming, still very much learning this stack. 
+
+**20 Feb 21**  
+Returning to this after a bit away on my CoinPurse project, need to re-factor a bunch to match that structure better. 
+
+
