+++ web/routes/home.py
@@ -1,5 +1,5 @@
 from flask import Blueprint, render_template
-from flask_login import current_user, login_required
+# from flask_login import current_user, login_required
 
 
 # Blueprint Configuration
@@ -7,7 +7,8 @@ home_bp = Blueprint('home_bp', __name__, template_folder='templates', static_fol
 
 
 @home_bp.route('/', methods=['GET'])
-@login_required
+# @login_required
 def index():
     """ Homepage """
-    return render_template('home/home.html', user=current_user.firstname)
+    # return render_template('home/home.html', user=current_user.firstname)
+    return render_template('home/home.html', user='<not set>')
