+++ web/templates/party/party_list.html
