+++ web/templates/player/player_add.html
