+++ web/templates/player/player_edit.html
