+++ web/templates/player/player_list.html
