+++ wsgi.py
@@ -1,0 +1,7 @@
+from web import create_app
+
+app = create_app()
+
+if __name__ == '__main__':
+    app.run(debug=True, host='0.0.0.0')
+
